requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==2.4.1
attrs==25.1.0
basketball_reference_web_scraper==4.15.3
bcrypt==4.2.1
blinker==1.9.0
bs4==0.0.2
cairocffi==1.7.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
coverage==7.6.10
chromadb>=0.4.1
cssselect2==0.8.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imblearn==0.0
ipykernel==6.29.5
itsdangerous==2.2.0
jedi==0.19.1
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lxml==5.3.0
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
narwhals==1.27.1
nbformat==5.10.4
networkx==3.4.2
numpy==2.1.3
openai==1.63.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.0
ptyprocess==0.7.0
python-dateutil==2.9.0.post0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydyf==0.11.0
pyparsing==3.2.1
pyphen==0.17.2
PyYAML>=6.0
pytz==2024.2
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
scipy==1.15.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
streamlit==1.44.1
sympy==1.13.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
torch==2.6.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wandb>=0.6.1
wcwidth==0.2.13
webencodings==0.5.1
whisper==1.1.10
xgboost==2.1.4
yarl==1.18.3
zopfli==0.2.3.post1
